Adicionados metodos de validação e para ações dentro das classes

+++ src/main/java/gabrielAMS/ex_campeonato/tabela_pont/repository/TabelaPontRepository.java
@@ -1,8 +1,8 @@
 package gabrielAMS.ex_campeonato.tabela_pont.repository;
 
-import gabrielAMS.ex_campeonato.tabela_pont.dto.DtoTabela;
+import gabrielAMS.ex_campeonato.tabela_pont.domain.DtoTabela;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TabelaPontRepository extends JpaRepository<DtoTabela, Integer> {
+public interface TabelaPontRepository extends JpaRepository<DtoTabela, Long> {
 
 }
